requirements.txt
google-generativeai>=0.8.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pandas>=2.1.0
numpy>=1.24.0
python-levenshtein>=0.25.0
tqdm>=4.66.0
click>=8.1.7
streamlit>=1.28.0
plotly>=5.17.0
pytest>=7.4.0
pillow>=10.1.0
jsonschema>=4.20.0
ollama>=0.3.0
anthropic>=0.34.0
huggingface_hub>=0.20.0